Move JRuby's ext and loader script (#39)
The strscan.rb file is only needed by JRuby, which does not have an
extension file format that loads itself. As noted in #35 having this
file present on TruffleRuby prevents their built-in strscan.rb from
loading, so the file was moved to lib/jruby/strscan.rb with lib/jruby
added as a require path. However this results in both lib/strscan.rb
and lib/jruby/strscan.rb getting copied into JRuby's stdlib when the
default gem is installed, leading to the confusing situation in #38.

This PR moves the jruby-specific strscan.rb under the jruby extension,
renaming that directory to be explicitly for JRuby, and adding that
new ext/jruby/lib path to the require paths so there's no nesting of
require paths to copy files multiple times.

* ext/java => ext/jruby
* lib/jruby/strscan.rb => ext/jruby/lib/strscan.rb
* Changes to Rakefile and gemspec for new locations

Fixes #38