Use JRuby 9.4.1.0 for CI due to region changes
The Region changes in JRuby's regex engine require that we support
both the old and new API for a while. Unfortunately the only
version of JRuby that contains both APIs is 9.4.1.0 so we have to
at least build against that version. Testing can occur against
any version, but for now this is the simplest way to get CI green.