Update Java ext version and add script logic for bump
The Java version is hardcoded like the C version, but has not been
getting updated. This will at least let it be updated by the rake
target version:bump, but any manual version number changes still
have to be done to both the C code and the Java code.

Fixes #97