Re-port new integer-scanning utility methods (#127)
GitHub: fix GH-125

This is a re-port of the C code from @byroot for fast base 10 and base
16 integer scanning.

In https://github.com/ruby/strscan/issues/125, @kou pointed out there's
an intermittent failure in the JRuby extension. We were unable to
confirm exactly the circumstances that cause that failure, but this
re-port should at least help reduce the change it is a bug in the
original Java code.